fixup! feat: add loaders for PostgreSQL catalog types
