--- psycopg/psycopg/types/cid.py
@@ -1,30 +1,0 @@
-from ..abc import AdaptContext
-from ._catalog import _Int4IntStrBinaryLoader, _IntStr, _StrSubclassLoader
-
-
-class CID(_IntStr):
-    __slots__ = ()
-
-    def _set_value(self, value: int) -> None:
-        if not 0 <= value < 2**32:
-            raise OverflowError("CID value must be in the unsigned 32 bits range")
-        self.value = value
-
-
-class CidLoader(_StrSubclassLoader[CID]):
-    """Load cid text values as `CID` (a string subclass)"""
-
-    cls = CID
-
-
-class CidBinaryLoader(_Int4IntStrBinaryLoader[CID]):
-    """Load cid binary values as `CID` (a string subclass)"""
-
-    cls = CID
-
-
-def register_default_adapters(context: AdaptContext) -> None:
-    adapters = context.adapters
-
-    adapters.register_loader("cid", CidLoader)
-    adapters.register_loader("cid", CidBinaryLoader)
